# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/workspace/create": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "create workspace",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workspace"
                ],
                "summary": "create workspace",
                "parameters": [
                    {
                        "type": "string",
                        "description": "workspace name",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.WorkspaceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/workspace/{workspace_id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "get workspace",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workspace"
                ],
                "summary": "get workspace",
                "parameters": [
                    {
                        "type": "string",
                        "description": "workspace id",
                        "name": "workspace_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.WorkspaceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/workspace/{workspace_id}/file-histories": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "get all files histories",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workspace"
                ],
                "summary": "get all files histories",
                "parameters": [
                    {
                        "type": "string",
                        "description": "workspace id",
                        "name": "workspace_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.FileHistoriesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "response.FileHistoriesResponse": {
            "type": "object",
            "properties": {
                "file_histories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.FileHistoryResponse"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/response.PaginationResponse"
                }
            }
        },
        "response.FileHistoryResponse": {
            "type": "object",
            "properties": {
                "files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.FileResponse"
                    }
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "response.FileResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "download_url": {
                    "type": "string"
                },
                "filename": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                },
                "tag": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/response.ProfileResponse"
                }
            }
        },
...
        "response.PaginationResponse": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer"
                },
                "next_page": {
                    "type": "integer"
                },
                "prev_page": {
                    "type": "integer"
                },
                "total_entities": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "response.ProfileResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "response.WorkspaceResponse": {
            "type": "object",
            "properties": {
                "creator": {
                    "$ref": "#/definitions/response.ProfileResponse"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.ProfileResponse"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}